refactor(domain): 도메인 모델을 도메인 패키지로 이동
Calculator, Operand 등 도메인 모델들을 도메인 패키지로 이동시킨다.
유지보수성을 위해 도메인 패키지를 만들어 도메인 로직을 관리한다.
