+++ README.md
@@ -10,10 +10,10 @@
         - [x]  문자열이 잘못된 값일 경우 예외를 발생시킨 후 어플리케이션을 종료한다.
         - [x]  구분자를 기준으로 분리한 각 숫자를 합한다.
     - 상태
-        - [ ]  기본 구분자
+        - [x]  기본 구분자
             - 쉼표(,)와 콜론(;)이다.
         - [ ]  커스텀 구분자
-        - [ ]  구분한 숫자 값
+        - [x]  구분한 숫자 값
 
 
 - **입출력 기능**
